StreetMatcher inserts first states with error instead of 0
When states were inserted with 0 it happened that worse roads (based on
error) were lengthened and then chose. (Road with error 14 instead of
road with error 0.02) Since always the road with lowest error is
chosen and the first time all roads have error 0 1st is chosen.
Changing first state error to to error fixes this.