Replaced i18next with JED translation library
I didn't removed i18next for now.
JED fixes all problems with plurals in French and Chinese files.
It also uses Plural forms from gettext itself and is meant to be used
with gettext.

The bigest problem is that it doesn't have language detection.
Currently i18next language detection is used which is a hack.