Add annotation for very long ways with tagged traffic signals
Sometimes when tagging incoming ways for traffic signals roads should
actually be split so that location of signals is more precise. But this
needs to be fixed in OSM data. And sometimes node should be tagged in
intersection if it is just a traffic signal for pedestrian
crossing. But because pavements aren't tagged there is no intersection
and it is interpreted as tagging a way. Since whole way can be couple
of kms long traffic signals costs on turns on this way are wrong.