Minorly improve WalkableAreaBuilder
Currently if starting nodes (nodes which also appear in ways, other
areas or are public transit stops) in areas had edges to other nodes
 which were a part of area ring. And all the other edges they have were
not inside current point visibility polygon those nodes vere not added
to starting vertices which is used in are edge prunning.

So it could happen that part of an area wasn't connected to the rest of
the graph and it was impossible to route over area.

This commit fixes this problem so that nodes are added to starting
vertices if they are starting nodes and pair of nodes is part of an
area ring.

Fixes #2111