Add support for transitNetworks in normal way of starting OTP
Added is --transitNetworks CLI parameter which can be used during
--build to build transitNetworks or during --server which loads
transitNetworks and makes Web API availible under otp_networks.

For now old Graph is always build together with transitNetworks because
it is usefull for comparisions.

During transitNetworks build envelope is made from all OSM vertices
coordinates that are in VertexStore. This is then returned in router
API point and will be used in the future in the clients. Center from
transitLayer is also returned in router.

TODO:
- multiple routers support
- coordinate envelope should be also filled with stop locations
- transitModes support
- convex Hull calculation support (it is used in analyst)